better usage tests
