--- internal/pnpm-test/test.sh
@@ -1,4 +1,0 @@
-pnpm install &&
-pnpm run build &&
-node bin/index.js &&
-echo "Live Share package was successfully imported and used by app using pnpm"
