+++ internal/usage-test/.gitignore
@@ -1,0 +1,3 @@
+# lock files not needed for usage tests
+pnpm-lock.yaml
+package-lock.json
