+++ internal/usage-test/cjs-test/.npmrc
