+++ internal/usage-test/esm-test/package.json
@@ -1,12 +1,11 @@
 {
-    "name": "live-share-pnpm-test",
+    "name": "live-share-esm-test",
     "version": "1.0.0",
     "description": "",
-    "main": "index.js",
+    "main": "src/index.js",
     "type": "module",
     "scripts": {
-        "test": "echo \"Error: no test specified\" && exit 1",
-        "build": "tsc -p tsconfig.json"
+        "test": "npm install && node src/index.js"
     },
     "keywords": [],
     "author": "",
@@ -16,7 +15,6 @@
         "fluid-framework": "2.1.0"
     },
     "devDependencies": {
-        "@fluidframework/test-runtime-utils": "^2.1.0",
-        "@types/node": "^20.14.14"
+        "@fluidframework/test-runtime-utils": "^2.1.0"
     }
 }
