+++ internal/usage-test/pnpm-test/package.json
@@ -1,0 +1,22 @@
+{
+    "name": "live-share-pnpm-test",
+    "version": "1.0.0",
+    "description": "",
+    "main": "bin/index.js",
+    "type": "module",
+    "scripts": {
+        "test": "pnpm install && pnpm run build && node bin/index.js",
+        "build": "tsc -p tsconfig.json"
+    },
+    "keywords": [],
+    "author": "",
+    "license": "ISC",
+    "dependencies": {
+        "@microsoft/live-share": "file:/../../packages/live-share",
+        "fluid-framework": "2.1.0"
+    },
+    "devDependencies": {
+        "@fluidframework/test-runtime-utils": "^2.1.0",
+        "@types/node": "^20.14.14"
+    }
+}
